[AC-4764] Create Methods Around "Became Entrant"
This commit laid the foundation for the inclusion of a "became entrant" entry in the JSON for organization history.  Unfortunately, despite each step working in the shell, I was unable to get the code working in the browser.  Additionally, I realized that I am not sure how to debug api code.  I tried pdb, but it just hung up.  I also realized that I am not sure how to update json with a list of items.  Seemed like I could pass in a list, but since it's currently null, I can't tell if that would work or not.  Also not entirely clear if we want a list of entrant strings under description.  Seems like that is information that should be more prominent?