[AC-4764] Move Short Name Option Over From Accelerate and Update Event Description
This commit updated the Event description by utilizing the short_name method from Accelerate.  As a result of the usage, it was necessary to change the JudgingRound and Program models.

This view is not quite finished.  I still need to add the earliest/latest conditions, but the functionality is there currently.