[AC-4900] Limit Input Strings For Expert Category
This commit cleans up the fields (mandatory or not) and adds a check to expert_category.  It is unclear where the error should be raised.  I think it makes more sense to raise it on the POST / PATCH, but I was unable to get a helper function similar to find_gender to work.  I have left the code in even though I don't think it belongs in ProfileHelper.