[AC-4900] Add user_type To ProfileHelper, Update Models, and Write Test
This commit moves the user_type field from the UserHelper options to the ProfileHelper.  Upon doing this and checking the API call, it became evident that user_type had not been moved over to this repo's Entrepreneur, Expert, or Member models.  That field was added here.  Finally a check on the inclusion of user_type in the GET was added.