[AC-5010] Convert URLs To Snake Case and General View Set Back To Camel Case
This commit changes the current CamelCase urls to snake_case and then changes them back for the model lookup in general_view_set.  There may well be a more elegeant and efficient way to do this.

The next step is excepting out the api calls that are not pure models.  There are currently 11 of those calls.