Build the directory into static/directory-dist
Why:

We have some unconventional behavior where were serving these assets
from `impact-api` but the domain that their hosted under is
`accelerate`'s. To do this, we have a nginx proxy in `accelerate` that
proxies any resquests to `/static/directory-dist/*` to `impact`. We
decided that to avoid confusion and possible naming collisions to change
the folder name from `dist` to `directory-dist`.

This PR:

Changes the docker files to cp the directory's `dist` folder to
`static/dist-directory` instead of `static/dist`